+++ web/src/app/layout.tsx
@@ -4,31 +4,10 @@ import "@/app/globals.css";
 import { StarBackground } from "@/components/effects/star-background";
 import { LanguageProvider } from "@/contexts/language-provider";
 import { SecureKeyProvider } from "@/contexts/secure-key-context";
-import { OfflineStatusToast } from "@/components/ui/offline-indicator";
-import { registerServiceWorker } from "@/lib/register-sw";
 import { AnalyticsProvider } from "@/components/ui/analytics-provider";
 import { ErrorBoundary } from "@/components/ui/error-boundary";
-import { PrivacyConsentBanner } from "@/components/ui/privacy-consent";
-import { Suspense, useEffect } from "react";
-import dynamic from "next/dynamic";
-import { initializeOptimizations } from "@/lib/performance/optimizations";
-
-// Lazy load non-critical components
-const LazyPrivacyConsentBanner = dynamic(
-  () => import("@/components/ui/privacy-consent").then(mod => ({ default: mod.PrivacyConsentBanner })),
-  {
-    loading: () => null,
-    ssr: false,
-  }
-);
-
-const LazyOfflineStatusToast = dynamic(
-  () => import("@/components/ui/offline-indicator").then(mod => ({ default: mod.OfflineStatusToast })),
-  {
-    loading: () => null,
-    ssr: false,
-  }
-);
+import { Suspense } from "react";
+import { ClientLayoutComponents } from "@/components/layout/client-layout";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -52,34 +31,6 @@ export const metadata: Metadata = {
   },
 };
 
-// Performance optimized client component
-function ClientOptimizations() {
-  useEffect(() => {
-    // Initialize performance optimizations
-    const initOptimizations = async () => {
-      try {
-        await initializeOptimizations({
-          enablePerformanceMonitoring: true,
-          enableCodeSplitting: true,
-          enableLazyLoading: true,
-          enableImageOptimization: true,
-          maxBundleSize: 250, // 250KB
-        });
-      } catch (error) {
-        console.warn('Failed to initialize performance optimizations:', error);
-      }
-    };
-
-    initOptimizations();
-
-    // Register service worker
-    if ('serviceWorker' in navigator) {
-      registerServiceWorker();
-    }
-  }, []);
-
-  return null;
-}
 
 export default function RootLayout({
   children,
@@ -123,12 +74,8 @@ export default function RootLayout({
                   </Suspense>
                 </main>
                 <Suspense fallback={null}>
-                  <LazyOfflineStatusToast />
+                  <ClientLayoutComponents />
                 </Suspense>
-                <Suspense fallback={null}>
-                  <LazyPrivacyConsentBanner />
-                </Suspense>
-                <ClientOptimizations />
               </LanguageProvider>
             </AnalyticsProvider>
           </SecureKeyProvider>
